Improve pin reading mechanism
Instead of requesting a pin read to the python layer and waiting for an
answer, we properly set input/ouput modes in the python layer and then
send a message whenever an input pin changes state. The C layer then
keeps a copy of those state and return them when asked.