Improve interrupt handling
We now stop the execution of the simulation while an interrupt is
running to avoid getting wrong values on MCU input pins while the
interrupt is running. It's not precisely the same situation as in the
real world where the interrupt *needs* to run fast enough to catch the
proper value, but we need this hack because of the delay we have during
inter-process communication.

Also, add IPC logging capabilitites. Very useful.