Reworked interrupts (again!)
They are now entirely managed on the C side. All we receive from the
simulation are pin changes and ticks.

Also, added a handy `icemu_send_debug()` function to send a debug value
to the simulation. We can only send values that fit in 5 bits, but
still, can be very very handy, especially when used with `debug_msgs_to`.