led: set fading counter during led_elapse()
We would previously set our fading counter right on pin change, when we
would go from "powered" to "unpowered" that had the unfortunate side
effect of lighting the LED in some situations when it should not, that
is, when pin changes that occur exactly at the same time (or with a very
fined tuned timing in our software) cause our LED to go on-then-off very
fast (or because of a wrong wire propagation order in the simulation) and
end up lighting our led for 10ms.

We now avoid this by setting our fading counter exclusively during
`led_elapse()`. That has the effect of possibly not simulating a glitch
that will occur on real hardware (that is, LED being very briefly lit
when they shouldn't), but at least it doesn't simulate another class of
glitches that will *not* happen. I prefer to avoid that latter type of
glitches.