mcu: fix elapse() return value
The return value for `mcu_elapse()` would be wrong in cases where:

1. Our next coming event is a runloop (not an interrupt or timer)
2. We haven't *just* ran a runloop

Now, in this case, we properly return the number of usecs left before we
need to run our next runloop. We would previously return the time until
the next interrupt, timer or `ICE_MAX_CHIP_ELAPSE` even in cases where
the next runloop would come before that.